C++: Expose a type resolution issue.

+++ cpp/ql/test/query-tests/Likely Bugs/Format/WrongTypeFormatArguments/Buildless/second.cpp
@@ -2,7 +2,7 @@
 
 int printf(const char * format, ...);
 
-// defines type `myFunctionPointerType`
-typedef int (*myFunctionPointerType) ();
+// defines type `myFunctionPointerType`, referencing `size_t`
+typedef size_t (*myFunctionPointerType) ();
 
 #include "include_twice.h"
